examples/chatbot-agent/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0,<3.0.0"]

[project]
authors = [
  {name = "Tanima Hossain"}
]
dependencies = [
  "fastapi (==0.116.0)",
  "chromadb (>=1.0.15,<2.0.0)",
  "python-dotenv (>=1.1.1,<2.0.0)",
  "requests (>=2.32.4,<3.0.0)",
  "streamlit (>=1.46.1,<2.0.0)",
  "pydantic (>=2.11.7,<3.0.0)",
  "uvicorn[standard] (>=0.35.0,<0.36.0)",
  "websockets (>=15.0.1,<16.0.0)"
]
description = "\"Chatbot using graphbit that stores conversation context in vectordb\""
name = "chatbot-agent"
readme = "README.md"
requires-python = ">=3.11"
version = "0.1.0"
